feat: remove wrapping logic over jina.math.evaluation
